+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 2021-10-25
+  - Track dispatcher that sent a message, or campaign that it belongs to and display in the messaging page
 # 2021-10-04
   - Redesigned campaigns list
   - During dispatching: search tasks via address